BestExperienceMessageModal: replaced logic from HFUI
